IN-236 refactor: put login logic into a hook
